// src/mgmt/interfaces.ts
export interface IMasterKey {
    secret: any;
    seed: string;
}

export interface IMasterKeyEncrypted {
    nonce: string;
    save: string;
}

export interface IKeypair {
    secretKey: Uint8Array;
    publicKey: Uint8Array;
    version: number | null;
}